authorize de medicalCenter, pacientes y workSchedules

+++ Backend/src/API-Turnos/WebApplication1/Controllers/PatientController.cs
@@ -4,11 +4,15 @@ using Application.Interfaces;
 using Domain.Entities;
 using Application.Models.Requests;
 using Application;
+using Microsoft.AspNetCore.Authorization;
+using Domain.Entities;
+using System.Security.Claims;
 
 namespace API.Controllers;
 
 [Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class PatientController : ControllerBase
 {   
     private readonly IPatientService _patientService;
@@ -33,10 +37,15 @@ public class PatientController : ControllerBase
     [HttpGet]
     public ActionResult<List<PatientDto>> GetAll()
     {
+        var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+        if (userRole != typeof(AdminMC).Name && userRole != typeof(SysAdmin).Name)
+            return Forbid();
         return _patientService.GetAll();
     }
 
     [HttpPost]
+    [AllowAnonymous]
     public IActionResult Create(PatientCreateRequest patientCreateRequest)
     {
         return Ok(_patientService.Create(patientCreateRequest));
@@ -45,6 +54,11 @@ public class PatientController : ControllerBase
     [HttpDelete("{id}")]
     public IActionResult Delete(int id)
     {
+        var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+        if (userRole != typeof(AdminMC).Name && userRole != typeof(SysAdmin).Name)
+            return Forbid();
+
         try
         {
             _patientService.Delete(id);
@@ -59,6 +73,11 @@ public class PatientController : ControllerBase
     [HttpPut("{id}")]
     public IActionResult Update(int id, PatientUpdateRequest patientUpdateRequest)
     {
+        var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+        if (userRole != typeof(AdminMC).Name && userRole != typeof(SysAdmin).Name)
+            return Forbid();
+
         try
         {
             _patientService.Update(id, patientUpdateRequest);
